Split dmdevice into dmnames and thinids
ThinIds are obviously associated with a pool, so should stay in thinpool,
but it turns out that we need names for all sorts of dm devices.
