+++ src/engine/strat_engine/thinpool/mod.rs
@@ -2,9 +2,9 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-mod dmdevice;
 mod filesystem;
 mod mdv;
+mod thinids;
 #[allow(module_inception)]
 mod thinpool;
 mod util;
